backend/requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pandas>=2.0.0
requests>=2.31.0
colorama>=0.4.6
redis>=2.0.0
ccxt>=4.0.0

torch>=2.0.0
scikit-learn>=1.3.0
stable-baselines3[extra]>=2.0.0
gymnasium>=0.29.0
shimmy>=1.0.0
